+++ src/dehydrate.js
@@ -3,7 +3,7 @@
 import isPrimitive from 'is-primitive'
 import { TYPE_AS_IS } from './common'
 
-function traverseAndCopy(chunks: Array<string>, src: Object, dest: any) {
+function traverseAndCopy(serializeValue: Function, chunks: Array<string>, src: Object, dest: any) {
   const chunk = chunks.shift()
 
   const keys = chunk ? [chunk] : Object.keys(src)
@@ -15,41 +15,47 @@ function traverseAndCopy(chunks: Array<string>, src: Object, dest: any) {
     }
 
     let item
-    if (isPrimitive(value)) {
+    const serializedValue = serializeValue(value)
+    if (isPrimitive(serializedValue)) {
       item = {
         type: TYPE_AS_IS,
-        value,
+        value: serializedValue,
       }
     } else {
-      let jsValue = value
-      const jsValueRef = value.toJS ? value.toJS() : value
-      if (Array.isArray(jsValueRef) && jsValue.toArray) {
-        jsValue = jsValue.toArray()
-      } else if (jsValue.toObject) {
-        jsValue = jsValue.toObject()
-      } else {
-        jsValue = jsValueRef
-      }
-
       item = {
         type: value.constructor.name,
-        value: Array.isArray(jsValue) ? [] : {},
+        value: Array.isArray(serializedValue) ? [] : {},
       }
 
-      traverseAndCopy(chunks, jsValue, item.value)
+      traverseAndCopy(serializeValue, chunks, serializedValue, item.value)
     }
 
     dest[key] = item
   })
 }
 
-export default function dehydrate(obj: Object, givenKeys: Array<string>) {
+export function defaultSerializeValue(value: any) {
+  if (!value) return value
+
+  let jsValue = value
+  const jsValueRef = value.toJS ? value.toJS() : value
+  if (Array.isArray(jsValueRef) && jsValue.toArray) {
+    jsValue = jsValue.toArray()
+  } else if (jsValue.toObject) {
+    jsValue = jsValue.toObject()
+  } else {
+    jsValue = jsValueRef
+  }
+  return jsValue
+}
+
+export default function dehydrate(serializeValue: Function, obj: Object, givenKeys: Array<string>) {
   const stateToPersist = { type: TYPE_AS_IS, value: {} }
 
   const keys = givenKeys.length ? givenKeys : Object.keys(obj)
 
   keys.forEach(function(key) {
-    traverseAndCopy(key.split('.'), obj, stateToPersist.value)
+    traverseAndCopy(serializeValue, key.split('.'), obj, stateToPersist.value)
   })
 
   return stateToPersist